Calculate check digit if not passed
Use the same calling conventions as the underlying GD::Barcode:: modules, which if given a code
one digit shorter than required for the format, will assume it's the check digit and calculate it.
Retrieve the full calculated text from the underlying GD::Barcode:: module and use that for the text
rendered underneath the bars.